pyproject.toml
[project]
name = "playground-github"
version = "0.0.1"
requires-python = ">=3.13.0"
dependencies = [
    "requests",
    "python-dotenv",
    "pygithub>=2.8.1",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]

[tool.pytest.ini_options]
pythonpath = "src"

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
